refactor: extract demo plan content to separate file
Moves the 334-line PLAN_CONTENT string from App.tsx to demoPlan.ts to reduce clutter.
